refactor(plugins/home): Implementations depend only on StorageApi
This patch removes the VisitsApiFactory, promotes CoreStorageVisitsApi
to the main implementation and LocalStorageVisitsApi starts to rely
on WebStorage.
