+++ plugins/home/src/api/index.ts
@@ -17,4 +17,3 @@
 export * from './CoreStorageVisitsApi';
 export * from './LocalStorageVisitsApi';
 export * from './VisitsApi';
-export * from './VisitsApiFactory';
